Create playRound() function to play a single round
This functions takes the user's choice and the computer's choice, and
compares them to determine the winner (if any) of the round. A message
is shown on the console, and the scores are updated accordingly. The
functions that get the choices were updated so that they display the
selection made