Refactor getHumanChoice() and playRound()
The way the website works now made mandatory to refactor these functions
simultaneously. A round is played whenever the player presses a button,
defining which button was pressed is delegated to getHumanChoice(). The
content of some visual elements is updated according to what button was
pressed and what choice made the computer