Replace playGame() for findWinner()
The addition of buttons to the user interface made obsolete the playGame
function. The new function findWinner executes what was missing in
playRound and was contained in playGame, which is determining the winner
of the game and modifying some DOM nodes. Minor modifications and
additions were made in the styles sheet